took app out of index
